+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 -   #2396 : Add support for NumPy `out` arguments.
 -   #2269 : Add support for `dtype`, `axis`, `keepdims`, `initial` arguments of `numpy.sum`.
 -   #2271 : Add support for `axis`, `keepdims`, `ord` arguments of `numpy.linalg.norm`.
+-   #2272 : Add support for `numpy.cross` and `numpy.linalg.cross` for vectors of size 3.
 
 ### Fixed